migrate production-only figure production to brei
