+++ .brei_history
@@ -3,5 +3,9 @@
   "docs/src/fig/life.gif": null,
   "docs/src/fig/boundary_types.png": "1cf470d765bbfa03ec1f8bec4e395959",
   "data/bs92-sealevel-curve.csv": null,
-  "docs/src/fig/bs92-fig8.svg": null
+  "docs/src/fig/bs92-fig8.svg": null,
+  "data/caps-osc.h5": null,
+  "docs/src/fig/b13-capsosc-crosssection.png": null,
+  "data/ca-prod-slope.h5": null,
+  "docs/src/fig/b13-crosssection.png": null
 }
