+++ .entangled/tasks.json
@@ -72,12 +72,72 @@
       "path": null
     },
     {
+      "description": null,
+      "creates": [
+        "data/caps-osc.h5"
+      ],
+      "requires": [
+        "src/CaProd.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/caps-osc.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "docs/src/fig/b13-capsosc-crosssection.png"
+      ],
+      "requires": [
+        "data/caps-osc.h5",
+        "ext/VisualizationExt.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/plot-caps-osc.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "data/ca-prod-slope.h5"
+      ],
+      "requires": [
+        "src/CaProd.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/production-only/cap-slope.jl"
+    },
+    {
+      "description": null,
+      "creates": [
+        "docs/src/fig/b13-crosssection.png"
+      ],
+      "requires": [
+        "data/ca-prod-slope.h5",
+        "ext/VisualizationExt.jl"
+      ],
+      "runner": "julia",
+      "stdout": null,
+      "stdin": null,
+      "script": null,
+      "path": "examples/production-only/plot-cap-slope.jl"
+    },
+    {
       "name": "figures",
       "requires": [
         "docs/src/fig/bs92-fig8.svg",
         "docs/src/fig/eca.png",
         "docs/src/fig/life.gif",
-        "docs/src/fig/boundary_types.png"
+        "docs/src/fig/boundary_types.png",
+        "docs/src/fig/b13-capsosc-crosssection.png",
+        "docs/src/fig/b13-crosssection.png"
       ]
     }
   ]
