+++ Project.toml
@@ -6,3 +6,11 @@ version = "0.1.0"
 [deps]
 HDF5 = "f67ccb44-e63f-5c2f-98bd-6dc0ccc4ba2f"
 Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999"
+
+[weakdeps]
+CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+GeometryBasics = "5c1252a2-5f33-56bf-86c9-59e7332b4326"
+
+[extensions]
+VisualizationExt = ["GeometryBasics", "CairoMakie"]
+
