--- examples/plot-cap-slope.jl
@@ -1,14 +1,0 @@
-# ~/~ begin <<docs/src/ca-with-production.md#examples/plot-cap-slope.jl>>[init]
-module Script
-using CarboKitten.Visualization
-using GLMakie
-
-function main()
-    f = Figure()
-    plot_crosssection(f[1, 1], "data/ca-prod-slope.h5")
-    save("docs/src/fig/b13-crosssection.png", f)
-end
-end
-
-Script.main()
-# ~/~ end
