+++ src/CarboKitten.jl
@@ -5,6 +5,6 @@ include("./Burgess2013.jl")
 include("./Utility.jl")
 # include("./BS92.jl")
 include("./CaProd.jl")
-# include("./Visualization.jl")
+include("./Visualization.jl")
 
 end # module CarboKitten
